Refactor face rec learning into package
